Added Product Activity and linked it to the search results
When barcode is scanned, results appear in the listView. From there a user can select the item and proceed to detailed search.
When the deailed search is finished, user can click once again on the list to trigger the product activity.
In the product activity user will see on a map markers with stores that offer the product

TODO: Implement Settings page